refactor: move appearance into a feature package
The layer packages sort code by what a class is, not by what it is
about. Everything appearance means sat spread over four of them, so
reading the area meant opening every layer, and changing it meant
touching every layer for one idea.

Appearance now lives in one directory that holds its own types, the
services it owns, and its own bindings. Settings and resources stay
horizontal, since appearance is only one of their callers. ADR 0009
writes down the shape so the next feature package can copy it.
