+++ mpvqc/appearance/services/__init__.py
@@ -1,0 +1,3 @@
+# SPDX-FileCopyrightText: mpvQC developers
+#
+# SPDX-License-Identifier: GPL-3.0-or-later
