+++ pyproject.toml
@@ -126,7 +126,12 @@ files = [
     "i18n/pt-PT.ts",
     "main.py",
     "mpvqc/__init__.py",
-    "mpvqc/appearance.py",
+    "mpvqc/appearance/__init__.py",
+    "mpvqc/appearance/domain.py",
+    "mpvqc/appearance/injections.py",
+    "mpvqc/appearance/services/__init__.py",
+    "mpvqc/appearance/services/color_scheme.py",
+    "mpvqc/appearance/services/palette_catalog.py",
     "mpvqc/application.py",
     "mpvqc/build.py",
     "mpvqc/close_event_filter.py",
@@ -172,7 +177,6 @@ files = [
     "mpvqc/services/__init__.py",
     "mpvqc/services/application_paths.py",
     "mpvqc/services/build_info.py",
-    "mpvqc/services/color_scheme.py",
     "mpvqc/services/comment_type_validator.py",
     "mpvqc/services/comment_types_policy.py",
     "mpvqc/services/comments/__init__.py",
@@ -216,7 +220,6 @@ files = [
     "mpvqc/services/label_width_calculator.py",
     "mpvqc/services/main_window.py",
     "mpvqc/services/mimetype_provider.py",
-    "mpvqc/services/palette_catalog.py",
     "mpvqc/services/platform/__init__.py",
     "mpvqc/services/platform/backend.py",
     "mpvqc/services/platform/embedded_player.py",
