+++ test/appearance/services/test_color_scheme.py
@@ -5,8 +5,9 @@
 import inject
 import pytest
 
-from mpvqc.appearance import Dark, FollowSystem, Light, Unknown
-from mpvqc.services import ColorSchemeService, SettingsService
+from mpvqc.appearance.domain import Dark, FollowSystem, Light, Unknown
+from mpvqc.appearance.services.color_scheme import ColorSchemeService
+from mpvqc.services import SettingsService
 
 SYSTEM = FollowSystem()
 LIGHT = Light()
