+++ test/models/test_accent_colors.py
@@ -6,9 +6,10 @@ import inject
 import pytest
 from PySide6.QtTest import QAbstractItemModelTester
 
-from mpvqc.appearance import ColorSchemePreference, Dark, FollowSystem, Light
+from mpvqc.appearance.domain import ColorSchemePreference, Dark, FollowSystem, Light
+from mpvqc.appearance.services.palette_catalog import PaletteCatalogService
 from mpvqc.models import MpvqcAccentColorModel
-from mpvqc.services import PaletteCatalogService, ResourceService
+from mpvqc.services import ResourceService
 
 SYSTEM = FollowSystem()
 LIGHT = Light()
